update mqtt and controls

+++ ESP32S3cam_controller/src/BasicController.cpp
@@ -3,14 +3,14 @@
 BasicController::BasicController(ServoMotor * joint, MotionProfile * profile): joint(joint), profile(profile) {}
 
 void BasicController::update() {
-    current_angle = (uint8_t) profile->update(target_angle);
+    current_angle = (float) profile->update(target_angle);
     joint->setAngle(current_angle);
 }
 
-void BasicController::moveTo(uint8_t angle) {
+void BasicController::moveTo(float angle) {
     target_angle = angle;
 }
 
-uint8_t BasicController::getCurrentAngle() {
+float BasicController::getCurrentAngle() {
     return(current_angle);
 }
