+++ ESP32S3cam_controller/src/BasicController.h
@@ -7,12 +7,12 @@ class BasicController {
         BasicController(ServoMotor * joint, MotionProfile * profile);
         ~BasicController();
         void update();
-        void moveTo(uint8_t angle);
-        uint8_t getCurrentAngle();
+        void moveTo(float angle);
+        float getCurrentAngle();
 
     private:
         ServoMotor * joint;
         MotionProfile * profile;
-        uint8_t target_angle;
-        uint8_t current_angle;
+        float target_angle;
+        float current_angle;
 };
