+++ ESP32S3cam_controller/src/Motor/ServoMotor.cpp
@@ -6,7 +6,7 @@ ServoMotor::ServoMotor(uint8_t id, uint16_t t_pulse_min, uint16_t t_pulse_max):
 
 ServoMotor::~ServoMotor(){};
 
-void ServoMotor::setAngle(uint8_t angle) {
+void ServoMotor::setAngle(float angle) {
     uint16_t pulselen = map(angle, 0, 180, t_pulse_min, t_pulse_max);
     PWMDriver::getInstance()->setPWM(id, pulselen);
 };
